+++ mzLib/Test/Omics/BioPolymerGroupTests/BioPolymerGroupTests.cs
@@ -1,15 +1,15 @@
-﻿using MassSpectrometry;
+﻿using System;
+using System.Collections.Generic;
+using System.Diagnostics.CodeAnalysis;
+using System.Linq;
+using MassSpectrometry;
 using NUnit.Framework;
 using Omics;
 using Omics.BioPolymerGroup;
 using Omics.Modifications;
 using Omics.SpectralMatch;
-using System;
-using System.Collections.Generic;
-using System.Diagnostics.CodeAnalysis;
-using System.Linq;
 
-namespace Test.Omics
+namespace Test.Omics.BioPolymerGroupTests
 {
     /// <summary>
     /// Unit tests for BioPolymerGroup class.
